pyproject.toml
[project]
name = "mongodb-rag-agent"
version = "0.1.0"
description = "Agentic RAG system with MongoDB Atlas Vector Search"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pydantic-ai>=0.1.0",
    "pymongo>=4.10.0",
    "openai>=1.58.0",
    "docling>=2.14.0",
    "docling-core>=2.4.0",
    "transformers>=4.47.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.1",
    "aiofiles>=24.1.0",
    "openai-whisper>=20240930",
    "slack-sdk>=3.27.0",
    "slack-bolt>=1.18.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
]
